fix quelques petits bug du changement
refs #1082 @0h15

+++ trunk/Commun/Utilitaire/CEcritureFichierBinaire.cpp
@@ -97,7 +97,7 @@ CEcritureFichierBinaire& operator< ( CEcritureFichierBinaire& out, const float& 
 ////////////////////////////////////////////////////////////////////////
 CEcritureFichierBinaire& operator< ( CEcritureFichierBinaire& out, const double& x )
 {
-   out.write( ( char* ) &x, sizeof ( float ) );
+   out.write( ( char* ) &x, sizeof ( double ) );
    return out;
 }
